+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add GgRanges.minInt, maxInt, isInRange
+
 ## [1.3.0] - 2024-04-29
 
 ### Added
@@ -65,6 +71,7 @@
 
 - Initial version.
 
+[Unreleased]: https://github.com/inlavigo/gg_list/compare/1.3.0...HEAD
 [1.3.0]: https://github.com/inlavigo/gg_list/compare/1.2.7...1.3.0
 [1.2.7]: https://github.com/inlavigo/gg_list/compare/1.2.6...1.2.7
 [1.2.6]: https://github.com/inlavigo/gg_list/compare/1.2.5...1.2.6